Refactor: rename module to match model
