+++ src/model/factory/platformOverviewItemFactory.ts
